Moved DataSharingSettings to domain module
