+++ server/src/main/resources/hibernate-mappings/DataSharingSettings.hbm.xml
@@ -20,7 +20,7 @@
         "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
 
 <hibernate-mapping package="com.thoughtworks.go.domain" default-access="field" default-lazy="false">
-    <class name="com.thoughtworks.go.domain.DataSharingSettings" table="DataSharingSettings">
+    <class name="com.thoughtworks.go.server.domain.DataSharingSettings" table="DataSharingSettings">
         <cache usage="read-write"/>
         <id name="id" column="id">
             <generator class="identity"/>
